Adds private AWS S3 Maven repository (refs #4)
As MetaMapLite is not OSS, a copy was deployed to a private Maven
repository hosted at AWS S3. Access credentials need to be provided at
runtime when a local copy is not found.

To access it, a S3 Maven wagon is needed as an extension, declared in
the POM file.