Skips tests when models are not available (refs #4)
MetaMap models are not OSS, so we should skip MetaMap tests when they're
not available.

To do so, we add lazy instantiation to MetaMapLiteFacade (to avoid a
FileNotFoundException when models are not available) and add a static
method for checking the model settings.

We separate loading properties in its own method so that it can be
called in a static context before instantiation.